requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.20.0
sqlalchemy==2.0.4
asyncpg==0.27.0
alembic==1.9.3
python-dotenv==1.0.0
pydantic==1.10.5
slowapi==0.1.9
starlette==0.36.3
psycopg2==2.9.5
python-multipart==0.0.6
typing-extensions==4.9.0
greenlet==2.0.2
Mako==1.2.4
MarkupSafe==2.1.2
limits==3.13.0
redis==5.0.3
celery==5.3.6
httpx==0.23.3
pytest==7.2.1
pytest-asyncio==0.20.3
